Start implementation Close

+++ SimipleNetworkProtocol/src/SendThreadQueue.cpp
@@ -35,7 +35,7 @@ namespace {
     }
 } // namespace
 
-void SendThreadQueue::WriteThread(std::stop_token stop, UDPQueue& udpQueue, SendThreadQueue& /*sendThreadQueue*/, const Socket& socket, uint16_t index)
+void SendThreadQueue::WriteThread(std::stop_token stop, UDPQueue& udpQueue, SendThreadQueue& /*sendThreadQueue*/, const Socket& socket, size_t index)
 {
     SetThreadName("WriteThread");
