+++ SimipleNetworkProtocol/src/SendThreadQueue.hpp
@@ -13,7 +13,7 @@ class SendThreadQueue {
 public:
     SendThreadQueue();
 
-    static void WriteThread(std::stop_token stop, UDPQueue& udpQueue, SendThreadQueue& sendThreadQueue, const Socket& socket, uint16_t index);
+    static void WriteThread(std::stop_token stop, UDPQueue& udpQueue, SendThreadQueue& sendThreadQueue, const Socket& socket, size_t index);
 
 private:
     std::chrono::milliseconds _pauseTime;
